Modifying some, Refactoring some
