+++ python/PENDENTE PENDENTE PENDENTE.txt
@@ -1,2 +1,3 @@
 matrizes.py
 primechecker.py
+oslibrary.py
