+++ python/tutorsss/librs/fcts5.py
