+++ python/tutorsss/librs/oslibrary.py
@@ -1,1 +1,2 @@
 import os
+#needing to study
