+++ python/tutorsss/readwritefiles/OPEN THIS FOLDER AS A SINGLE WORKSPACE.txt
@@ -1,2 +1,2 @@
-open this not as 'pythona' in workspace, it wont work that way
-open it separetely(select only this folder)
+don't 'python' folder as a workspace, it wont work that way
+open it separetely (select only this current folder)
