--- python/tutorsss/readwritefiles/test.txt
@@ -1,4 +1,0 @@
-atila - 1
-nana - 2
-well - 3
-marx - 10
